Only parse function calls if the identifier is a function. Fixes #193
The function call syntax is ambiguous in certain MIPS opcodes such as
"lw a0,offset(a0)". Here, offset is intended to be used as a label
address, but was actually interpreted as a function call.

The ambiguity is resolved by only generating a function call if the
identifier actually refers to an existing function.